refactor: TemperatureRowを別ファイルで管理
